+++ backend/src/modules/boards/applications/update.board.application.ts
@@ -1,6 +1,4 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { ColumnDeleteCardsDto } from 'src/libs/dto/colum.deleteCards.dto';
-import { UpdateColumnDto } from '../dto/column/update-column.dto';
 import { UpdateBoardDto } from '../dto/update-board.dto';
 import { UpdateBoardApplicationInterface } from '../interfaces/applications/update.board.application.interface';
 import { UpdateBoardServiceInterface } from '../interfaces/services/update.board.service.interface';
@@ -19,12 +17,5 @@ export class UpdateBoardApplication implements UpdateBoardApplicationInterface {
 
 	mergeBoards(subBoardId: string, userId: string) {
 		return this.updateBoardService.mergeBoards(subBoardId, userId);
-	}
-
-	updateColumn(boardId: string, column: UpdateColumnDto) {
-		return this.updateBoardService.updateColumn(boardId, column);
-	}
-	deleteCardsFromColumn(boardId: string, column: ColumnDeleteCardsDto) {
-		return this.updateBoardService.deleteCardsFromColumn(boardId, column);
 	}
 }
