+++ backend/src/modules/boards/interfaces/services/update.board.service.interface.ts
@@ -1,7 +1,5 @@
 import { LeanDocument } from 'mongoose';
-import { ColumnDeleteCardsDto } from 'src/libs/dto/colum.deleteCards.dto';
 import { TeamDto } from 'src/modules/communication/dto/team.dto';
-import { UpdateColumnDto } from '../../dto/column/update-column.dto';
 import { UpdateBoardDto } from '../../dto/update-board.dto';
 import { BoardDocument } from '../../schemas/board.schema';
 
@@ -10,13 +8,4 @@ export interface UpdateBoardServiceInterface {
 
 	mergeBoards(subBoardId: string, userId: string): Promise<LeanDocument<BoardDocument> | null>;
 	updateChannelId(teams: TeamDto[]);
-	updateColumn(
-		boardId: string,
-		column: UpdateColumnDto
-	): Promise<LeanDocument<BoardDocument> | null>;
-
-	deleteCardsFromColumn(
-		boardId: string,
-		column: ColumnDeleteCardsDto
-	): Promise<LeanDocument<BoardDocument> | null>;
 }
