+++ backend/src/modules/boards/services/update.board.service.ts
@@ -15,21 +15,18 @@ import { CommunicationServiceInterface } from 'src/modules/communication/interfa
 import * as CommunicationsType from 'src/modules/communication/interfaces/types';
 import { GetTeamServiceInterface } from 'src/modules/teams/interfaces/services/get.team.service.interface';
 import * as Teams from 'src/modules/teams/interfaces/types';
-import * as Votes from 'src/modules/votes/interfaces/types';
+import * as Cards from 'src/modules/cards/interfaces/types';
 import User, { UserDocument } from 'src/modules/users/entities/user.schema';
 import { UpdateBoardDto } from '../dto/update-board.dto';
 import { ResponsibleType } from '../interfaces/responsible.interface';
 import { UpdateBoardServiceInterface } from '../interfaces/services/update.board.service.interface';
 import Board, { BoardDocument } from '../schemas/board.schema';
 import BoardUser, { BoardUserDocument } from '../schemas/board.user.schema';
-import { BoardDataPopulate } from '../utils/populate-board';
-import { UpdateColumnDto } from '../dto/column/update-column.dto';
 import { UPDATE_FAILED } from 'src/libs/exceptions/messages';
 import SocketGateway from 'src/modules/socket/gateway/socket.gateway';
-import { DeleteVoteServiceInterface } from 'src/modules/votes/interfaces/services/delete.vote.service.interface';
-import Column from '../schemas/column.schema';
-import ColumnDto from '../dto/column/column.dto';
-import { ColumnDeleteCardsDto } from 'src/libs/dto/colum.deleteCards.dto';
+import Column from '../../columns/schemas/column.schema';
+import ColumnDto from '../../columns/dto/column.dto';
+import { DeleteCardService } from 'src/modules/cards/interfaces/services/delete.card.service.interface';
 
 @Injectable()
 export default class UpdateBoardServiceImpl implements UpdateBoardServiceInterface {
@@ -42,8 +39,8 @@ export default class UpdateBoardServiceImpl implements UpdateBoardServiceInterfa
 		@InjectModel(BoardUser.name)
 		private boardUserModel: Model<BoardUserDocument>,
 		private socketService: SocketGateway,
-		@Inject(Votes.TYPES.services.DeleteVoteService)
-		private deleteVoteService: DeleteVoteServiceInterface
+		@Inject(Cards.TYPES.services.DeleteCardService)
+		private deleteCardService: DeleteCardService
 	) {}
 
 	/**
@@ -131,6 +128,10 @@ export default class UpdateBoardServiceImpl implements UpdateBoardServiceInterfa
 				await Promise.all(promises);
 			}
 
+			/**
+			 * TODO:
+			 * When the mainBoardId starts to be returned by the board, remove this query to the boardModel
+			 */
 			const mainBoardId = await this.boardModel
 				.findOne({ dividedBoards: { $in: boardId } })
 				.select('_id')
@@ -181,25 +182,7 @@ export default class UpdateBoardServiceImpl implements UpdateBoardServiceInterfa
 				return board.columns.find((column) => column._id.toString() === deletedColumnId)?.cards;
 			});
 
-			cardsToDelete.forEach((cards) => {
-				cards.items.forEach(async (card) => {
-					const votesByUser = new Map<string, number>();
-
-					card.votes.forEach((userId) => {
-						if (!votesByUser.has(userId.toString())) {
-							votesByUser.set(userId.toString(), 1);
-						} else {
-							const count = votesByUser.get(userId.toString());
-
-							votesByUser.set(userId.toString(), count + 1);
-						}
-					});
-
-					votesByUser.forEach(async (votesCount, userId) => {
-						await this.deleteVoteService.decrementVoteUser(board.id, userId, -votesCount);
-					});
-				});
-			});
+			await this.deleteCardService.deleteCardsFromColumn(boardId, cardsToDelete);
 		}
 
 		/**
@@ -439,90 +422,5 @@ export default class UpdateBoardServiceImpl implements UpdateBoardServiceInterfa
 				this.boardModel.updateOne({ _id: team.boardId }, { slackChannelId: team.channelId })
 			)
 		);
-	}
-
-	updateColumn(boardId: string, column: UpdateColumnDto) {
-		const board = this.boardModel
-			.findOneAndUpdate(
-				{
-					_id: boardId,
-					'columns._id': column._id
-				},
-				{
-					$set: {
-						'columns.$[column].color': column.color,
-						'columns.$[column].title': column.title,
-						'columns.$[column].cardText': column.cardText,
-						'columns.$[column].isDefaultText': column.isDefaultText
-					}
-				},
-				{
-					arrayFilters: [{ 'column._id': column._id }],
-					new: true
-				}
-			)
-			.populate(BoardDataPopulate)
-			.lean()
-			.exec();
-
-		if (!board) throw new BadRequestException(UPDATE_FAILED);
-
-		if (column.socketId) this.socketService.sendUpdatedBoard(boardId, column.socketId);
-
-		return board;
-	}
-
-	async deleteCardsFromColumn(boardId: string, column: ColumnDeleteCardsDto) {
-		const board = await this.boardModel.findById(boardId).exec();
-
-		if (!board) throw new BadRequestException(UPDATE_FAILED);
-
-		const columnUpdate = board.columns.find((col) => String(col._id) === column.id)?.cards;
-
-		columnUpdate.forEach((cards) => {
-			cards.items.forEach(async (card) => {
-				const votesByUser = new Map<string, number>();
-
-				card.votes.forEach((userId) => {
-					if (!votesByUser.has(userId.toString())) {
-						votesByUser.set(userId.toString(), 1);
-					} else {
-						const count = votesByUser.get(userId.toString());
-
-						votesByUser.set(userId.toString(), count + 1);
-					}
-				});
-
-				votesByUser.forEach(async (votesCount, userId) => {
-					await this.deleteVoteService.decrementVoteUser(boardId, userId, -votesCount);
-				});
-			});
-		});
-
-		const updateBoard = await this.boardModel
-			.findOneAndUpdate(
-				{
-					_id: boardId,
-					'columns._id': column.id
-				},
-				{
-					$set: {
-						'columns.$[column].cards': []
-					}
-				},
-				{
-					arrayFilters: [{ 'column._id': column.id }],
-					new: true
-				}
-			)
-			.populate(BoardDataPopulate)
-			.lean()
-			.exec();
-
-		if (!updateBoard) throw new BadRequestException(UPDATE_FAILED);
-
-		if (column.socketId) this.socketService.sendUpdatedBoard(boardId, column.socketId);
-
-		return updateBoard;
 	}
 }
