+++ backend/src/modules/columns/columns.providers.ts
@@ -1,0 +1,13 @@
+import { UpdateColumnApplicationImpl } from './applications/update.columns.application';
+import { TYPES } from './interfaces/types';
+import UpdateColumnServiceImpl from './services/update.column.service';
+
+export const updateColumnService = {
+	provide: TYPES.services.UpdateColumnService,
+	useClass: UpdateColumnServiceImpl
+};
+
+export const updateColumnApplication = {
+	provide: TYPES.applications.UpdateColumnApplication,
+	useClass: UpdateColumnApplicationImpl
+};
