+++ backend/src/modules/columns/dto/colum.deleteCards.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString } from 'class-validator';
-import { BaseDto } from './base.dto';
+import { BaseDto } from '../../../libs/dto/base.dto';
 
 export class ColumnDeleteCardsDto extends BaseDto {
 	@ApiProperty()
