+++ backend/src/modules/columns/interfaces/types.ts
@@ -1,0 +1,8 @@
+export const TYPES = {
+	services: {
+		UpdateColumnService: 'UpdateColumnService'
+	},
+	applications: {
+		UpdateColumnApplication: 'UpdateColumnApplication'
+	}
+};
