+++ frontend/src/api/boardService.tsx
@@ -73,7 +73,7 @@ export const deleteBoardRequest = async ({
 export const updateColumnRequest = (
   columnData: ColumnType & { boardId: string },
 ): Promise<BoardType> =>
-  fetchData(`/boards/${columnData.boardId}/column/${columnData._id}`, {
+  fetchData(`/columns/${columnData.boardId}/column/${columnData._id}`, {
     method: 'PUT',
     data: columnData,
   });
@@ -81,7 +81,7 @@ export const updateColumnRequest = (
 export const deleteCardsFromColumnRequest = (
   columnData: ColumnDeleteCards & { boardId: string },
 ): Promise<BoardType> =>
-  fetchData(`/boards/${columnData.boardId}/column/${columnData.id}/cards`, {
+  fetchData(`/columns/${columnData.boardId}/column/${columnData.id}/cards`, {
     method: 'PUT',
     data: columnData,
   });
